Refactor plugin internals
