+++ gradle/plugins/src/main/kotlin/io/dotanuki/norris/gradle/AutoModulePlugin.kt
@@ -5,11 +5,11 @@ import io.dotanuki.norris.gradle.internal.ModuleConvention.ANDROID_APPLICATION
 import io.dotanuki.norris.gradle.internal.ModuleConvention.ANDROID_FEATURE_LIBRARY
 import io.dotanuki.norris.gradle.internal.ModuleConvention.ANDROID_PLATFORM_LIBRARY
 import io.dotanuki.norris.gradle.internal.ModuleConvention.KOTLIN_PLATFORM_LIBRARY
-import io.dotanuki.norris.gradle.internal.applyAndroidApplicationConventions
-import io.dotanuki.norris.gradle.internal.applyAndroidFeatureLibraryConventions
-import io.dotanuki.norris.gradle.internal.applyAndroidPlatformLibraryConventions
-import io.dotanuki.norris.gradle.internal.applyKotlinProjectConventions
-import io.dotanuki.norris.gradle.internal.applyTestLoggingConventions
+import io.dotanuki.norris.gradle.internal.conventions.applyAndroidApplicationConventions
+import io.dotanuki.norris.gradle.internal.conventions.applyAndroidFeatureLibraryConventions
+import io.dotanuki.norris.gradle.internal.conventions.applyAndroidPlatformLibraryConventions
+import io.dotanuki.norris.gradle.internal.conventions.applyKotlinProjectConventions
+import io.dotanuki.norris.gradle.internal.conventions.applyTestLoggingConventions
 import org.gradle.api.Plugin
 import org.gradle.api.Project
 
@@ -28,7 +28,6 @@ class AutoModulePlugin : Plugin<Project> {
                 ANDROID_PLATFORM_LIBRARY -> {
                     pluginManager.apply("kotlin-android")
                     pluginManager.apply("com.android.library")
-
                     applyKotlinProjectConventions()
                     applyAndroidPlatformLibraryConventions()
                 }
@@ -36,14 +35,12 @@ class AutoModulePlugin : Plugin<Project> {
                     pluginManager.apply("kotlin-android")
                     pluginManager.apply("com.android.library")
                     pluginManager.apply("com.dropbox.dropshots")
-
                     applyKotlinProjectConventions()
                     applyAndroidFeatureLibraryConventions()
                 }
                 ANDROID_APPLICATION -> {
                     pluginManager.apply("kotlin-android")
                     pluginManager.apply("com.android.application")
-
                     applyKotlinProjectConventions()
                     applyAndroidApplicationConventions()
                 }
