+++ gradle/plugins/src/main/kotlin/io/dotanuki/norris/gradle/internal/AppVersion.kt
@@ -1,0 +1,9 @@
+package io.dotanuki.norris.gradle.internal
+
+internal data class AppVersion(
+    val name: String,
+    val major: Int,
+    val minor: Int,
+    val patch: Int,
+    val code: Int
+)
