+++ gradle/plugins/src/main/kotlin/io/dotanuki/norris/gradle/internal/conventions/KotlinModuleConventions.kt
@@ -1,4 +1,4 @@
-package io.dotanuki.norris.gradle.internal
+package io.dotanuki.norris.gradle.internal.conventions
 
 import org.gradle.api.JavaVersion
 import org.gradle.api.Project
